requirements.txt
appdirs==1.4.4
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
autopep8==2.0.2
az-iranian-bank-gateways==1.9.0
cached-property==1.5.2
certifi==2023.7.22
cffi==1.16.0
channels==3.0.4
channels-redis==4.1.0
charset-normalizer==2.1.1
cmake==3.27.7
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
coverage==7.4.0
cryptography==41.0.5
daphne==3.0.2
defusedxml==0.7.1
Django==4.2.6
django-cors-headers==4.3.0
django-environ==0.11.2
django-filter==23.3
django-hashid-field==3.3.7
django-phonenumber-field==7.2.0
django-phonenumbers==1.0.1
django-redis==5.4.0
django-rest-swagger==2.2.0
django-zarinpal==1.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
drf-yasg==1.21.5
environ==1.0
environs==9.5.0
exceptiongroup==1.2.0
gunicorn==21.2.0
hashids==1.3.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
install==1.3.5
isodate==0.6.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
model-bakery==1.17.0
msgpack==1.0.7
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.2
phonenumbers==8.13.23
pika==1.3.2
Pillow==10.1.0
platformdirs==3.11.0
pluggy==1.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.19.0
PyJWT==2.8.0
pyOpenSSL==23.3.0
pypi-install==0.0.5
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.7.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.30.2
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rest-framework-simplejwt==0.0.2
rpds-py==0.10.6
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.8
service-identity==23.1.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.18
zeep==3.4.0
zope.interface==6.1